fix: convert formatter to typescript
