+++ uimodule/src/model/models.ts
@@ -1,0 +1,13 @@
+import JSONModel from "sap/ui/model/json/JSONModel";
+import BindingMode from "sap/ui/model/BindingMode";
+import * as Device from "sap/ui/Device";
+
+export default {
+
+	createDeviceModel : () => {
+		const oModel = new JSONModel(Device);
+		oModel.setDefaultBindingMode(BindingMode.OneWay);
+		return oModel;
+	}
+
+};
